lambda/requirements.txt
tensorflow-cpu==2.21.0
numpy
boto3
